requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.10
pyyaml>=6.0
jinja2>=3.1.0
python-dateutil>=2.8.2
shapely>=2.0.0
fake-useragent>=1.4.0
cloudscraper>=1.2.71
curl-cffi>=0.7.0
